Prefix mkincl targets with with mkincl
The generic targets 'init' and 'clean' are bound to experience conflicts
with user Makefiles. As such, rename the targets in .mkincl/init.mk to
'mkincl-init' and 'mkincl-clean'.

An example of where the past names were problematic: If a user
repository defines a target named 'clean' that does removes arbitrary
directory, this target will shadow mkincl's 'clean' target which will
break the 'init' target:

    $ make init
    Makefile:4: warning: overriding recipe for target 'clean'
    .mkincl/init.mk:15: warning: ignoring old recipe for target 'clean'
    echo hello
    hello
    -- Initializing provider .mkincl/inits/shell.sh
    . /home/carsme/repos/git.got.jeppesensystems.com/fpc/cloud/shfpxclient/.mkincl/inits/shell.sh \
    && mkdir -p .mkincl/providers \
    && git clone --quiet $URL .mkincl/providers/$NAME \
    && git -C .mkincl/providers/$NAME reset --quiet --hard $VERSION
    fatal: destination path '.mkincl/providers/shell' already exists and is not an empty directory.
    make: *** [.mkincl/init.mk:8: .mkincl/inits/shell.sh] Error 128